chore: remove concept of MessageItem, replace with MessageContext
