+++ discord_bot/cogs/music_helpers/message_context.py
@@ -1,6 +1,9 @@
+from datetime import datetime, timezone
 from uuid import uuid4
 
 from discord import Message
+
+from discord_bot.cogs.music_helpers.common import MessageLifecycleStage
 
 class MessageContext():
     '''
@@ -9,12 +12,16 @@ class MessageContext():
     def __init__(self, guild_id: int, channel_id: int):
         self.guild_id = guild_id
         self.channel_id = channel_id
-
         self.uuid = uuid4()
+        self.created_at = datetime.now(timezone.utc)
+        self.lifecycle_stage = MessageLifecycleStage.SEND
 
+        # Set after
         self.message_id = None
         self.message = None
-
+        self.message_content = None
+        self.delete_after = None
+        self.function = None
 
     def set_message(self, message: Message):
         '''
